requirements.txt
python-docx==1.1.0
docxtpl==0.16.7
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
num2words==0.5.13
Jinja2==3.1.2
PyYAML==6.0.1
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0


